Refactor css.
