+++ src/components/AllMovies/AllMovies.css
@@ -1,5 +1,4 @@
 .card-container {
   display: flex;
   flex-wrap: wrap;
-  background-color: rgb(5, 4, 4)
 }
