+++ src/components/App/App.css
@@ -1,3 +1,5 @@
 .App {
   text-align: center;
+  color: white;
+  background-color: black;
 }
