+++ src/components/MovieCard/MovieCard.css
@@ -18,12 +18,6 @@
   box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
 }
 
-h2 {
-  margin: 0;
-  color: white;
-}
-
 h3 {
   margin: 0;
-  color: white;
 }
